Add files via upload
Read Me
(The file provided is all of the written code in the project)

How to use:
Download ZIP, Unpack file, Type in Npm run dev in terminal and copy the server link into a browser to access the information. 

Items Used:
Temp and Humidity Measure:
The DHT22, or AM2302, is a popular digital humidity and temperature sensor frequently used with ESP32 microcontrollers in IoT applications. Operating on a single-wire digital interface and typically powered by 3.3V, it offers accurate readings over a temperature range of -40°C to 80°C and a humidity range of 0% to 100%. With a resolution of 0.1°C for temperature and 2% for humidity, it provides reliable environmental data. The sensor is known for its affordability, simplicity, and compatibility with various projects, making it a go-to choice for monitoring and controlling temperature and humidity in smart home systems, weather stations, and other connected devices.

Wires:
In wiring an ESP32 with a DHT22 sensor, three fundamental connections establish communication. The VCC (Voltage Common Collector) wire links to a 3.3V output on the ESP32, providing power. Ground (GND) is established by connecting the GND wire to an available ground pin on the ESP32, forming the electrical reference. The Data wire, pivotal for information exchange, is linked to any digital GPIO (General Purpose Input/Output) pin on the ESP32. It facilitates the bidirectional flow of data between the ESP32 and the DHT22 sensor. A crucial addition is a pull-up resistor (commonly 4.7kΩ) between the Data pin and the VCC pin, ensuring stable data transmission.